Migrate neovim to use lua (and associated wins)
This is an overhaul of most of my vim.

By leaning into lazy.nvim I can group my plugins along with their
related mappings, highlights and dependencies - allowing for all related
elements for a plugin being contained.

It also allows for a faster experience not only from using lua for a
variety of plugins (ie. nerdtree -> nvim-tree) but also as they are
loaded lazily (often by the exact keybinding used with them)